@@ -4,29 +4,26 @@ public partial class HistoricoPage : ContentPage
 {
     public IHistoricoService C_HistoricoService { get; set; }
     public ITreinoService C_TreinoService { get; set; }
-    public IExercicioService C_ExercicioService { get; set; }
 
     public HistoricoPage(
         IHistoricoService p_HistoricoService,
-        ITreinoService p_TreinoService,        
-        IExercicioService p_ExercicioService)
+        ITreinoService p_TreinoService)
 	{
 		InitializeComponent();
 
         C_HistoricoService = p_HistoricoService;
         C_TreinoService = p_TreinoService;
-        C_ExercicioService = p_ExercicioService;
-	}
+    }
 
     protected override async void OnAppearing()
     {
         base.OnAppearing();
-        lst_Historico.ItemsSource = await C_HistoricoService.CM_VerHistoricoAsync();
+        lst_Historico.ItemsSource = await C_HistoricoService.CM_VerHistoricoAsync(null, null);
     }
 
     private async void lst_Historico_ItemSelected(object sender, SelectedItemChangedEventArgs e)
     {
         var m_Treino = e.SelectedItem as Treino;
-        await Navigation.PushAsync(new TreinoPage(m_Treino));
+        await Shell.Current.Navigation.PushAsync(new TreinoPage(C_TreinoService, m_Treino.Id, m_Treino.Nome, true), true);
     }
 }
