@@ -2,5 +2,5 @@
 
 public interface IHistoricoService
 {
-    Task<IEnumerable<Treino>> CM_VerHistoricoAsync();
+    Task<IEnumerable<Treino>> CM_VerHistoricoAsync(DateTime? p_Inicio, DateTime? p_Fim);
 }
