@@ -2,21 +2,23 @@
 
 public class Exercicio
 {
-    public Exercicio() { }
-
-    public Exercicio(string p_Nome, string p_Repeticao, int p_Carga)
-    {
-        Nome = p_Nome;
-        Repeticao = p_Repeticao;
-        Carga = p_Carga;
-    }
-
-    public string Nome { get; set; }
-    public string Repeticao { get; set; }
+    public int Id { get; set; }
+    public string Descricao { get; set; }
 
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? Carga { get; set; }
+    public int? Series { get; set; }
+
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public int? Repeticao { get; set; }
+
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public double? Tempo { get; set; }
+
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public double? Carga { get; set; }
 
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public bool? Concluido { get; set; }
+
+    public int CodigoTreino { get; set; }
 }
