@@ -14,46 +14,46 @@ public class TreinoHelperTestes
         C_TreinoHelper = new TreinoHelper();
         C_TreinosDesordenados = new List<Treino>
         {
-            new Treino('A', "Teste1"),
-            new Treino('F', "Teste2"),
-            new Treino('B', "Teste3"),
-            new Treino('U', "Teste4")
+            new Treino("Teste1"),
+            new Treino("Teste2"),
+            new Treino("Teste3"),
+            new Treino("Teste4")
         };
         C_TodosTreinos = new List<Treino>
         {
-            new Treino('A', "Teste1"),
-            new Treino('B', "Teste2"),
-            new Treino('C', "Teste3"),
-            new Treino('D', "Teste4"),
-            new Treino('E', "Teste5"),
-            new Treino('F', "Teste6"),
-            new Treino('G', "Teste7"),
-            new Treino('H', "Teste8"),
-            new Treino('I', "Teste9"),
-            new Treino('J', "Teste10"),
-            new Treino('K', "Teste11"),
-            new Treino('L', "Teste12"),
-            new Treino('M', "Teste13"),
-            new Treino('N', "Teste14"),
-            new Treino('O', "Teste15"),
-            new Treino('P', "Teste16"),
-            new Treino('Q', "Teste17"),
-            new Treino('R', "Teste18"),
-            new Treino('S', "Teste19"),
-            new Treino('T', "Teste20"),
-            new Treino('U', "Teste21"),
-            new Treino('V', "Teste22"),
-            new Treino('W', "Teste23"),
-            new Treino('X', "Teste24"),
-            new Treino('Y', "Teste25"),
-            new Treino('Z', "Teste26")
+            new Treino("Teste1"),
+            new Treino("Teste2"),
+            new Treino("Teste3"),
+            new Treino("Teste4"),
+            new Treino("Teste5"),
+            new Treino("Teste6"),
+            new Treino("Teste7"),
+            new Treino("Teste8"),
+            new Treino("Teste9"),
+            new Treino("Teste10"),
+            new Treino("Teste11"),
+            new Treino("Teste12"),
+            new Treino("Teste13"),
+            new Treino("Teste14"),
+            new Treino("Teste15"),
+            new Treino("Teste16"),
+            new Treino("Teste17"),
+            new Treino("Teste18"),
+            new Treino("Teste19"),
+            new Treino("Teste20"),
+            new Treino("Teste21"),
+            new Treino("Teste22"),
+            new Treino("Teste23"),
+            new Treino("Teste24"),
+            new Treino("Teste25"),
+            new Treino("Teste26")
         };
         C_SemTreinos = new List<Treino>();
         C_SemTreinoA = new List<Treino>
         {
-            new Treino('F', "Teste1"),
-            new Treino('B', "Teste2"),
-            new Treino('U', "Teste3")
+            new Treino("Teste1"),
+            new Treino("Teste2"),
+            new Treino("Teste3")
         };
     }
 
